+++ utils/createOrders.ts
@@ -6,6 +6,22 @@ import { IexecLibOrders_v5 } from '../typechain';
 import constants from './constants';
 import { utils } from './odb-tools';
 
+interface Iexec<T> {
+    app: T;
+    dataset: T;
+    workerpool: T;
+    requester: T;
+}
+
+export interface IexecAccounts extends Iexec<SignerWithAddress> {}
+
+export interface IexecOrders extends Iexec<Record<string, any>> {
+    app: IexecLibOrders_v5.AppOrderStruct;
+    dataset: IexecLibOrders_v5.DatasetOrderStruct;
+    workerpool: IexecLibOrders_v5.WorkerpoolOrderStruct;
+    requester: IexecLibOrders_v5.RequestOrderStruct;
+}
+
 export function createEmptyAppOrder(): IexecLibOrders_v5.AppOrderStruct {
     return {
         app: constants.NULL.ADDRESS,
@@ -95,7 +111,20 @@ export function buildCompatibleOrders(
     requestOrder.tag = tag;
     datasetOrder.tag = tag;
     workerpoolOrder.tag = tag;
-    return { appOrder, datasetOrder, workerpoolOrder, requestOrder };
+
+    return {
+        orders: {
+            app: appOrder,
+            dataset: datasetOrder,
+            workerpool: workerpoolOrder,
+            requester: requestOrder,
+        } as IexecOrders,
+        // Expose orders differently to make them easier to use in tests
+        appOrder: appOrder,
+        datasetOrder: datasetOrder,
+        workerpoolOrder: workerpoolOrder,
+        requestOrder: requestOrder,
+    };
 }
 
 /**
@@ -113,6 +142,26 @@ export function buildDomain(domain?: TypedDataDomain | undefined) {
     }
     const domainSeparator = ethers.utils._TypedDataEncoder.hashDomain(domain);
     return { domain, domainSeparator };
+}
+
+/**
+ * Sign all orders required by `matchOrder` calls.
+ * @param domain typed data domain for EIP-712 signature
+ * @param orders orders to sign (app, dataset, workerpool and requester orders)
+ * @param signers accounts which will respectively sign orders according to their
+ *  role
+ */
+export async function signOrders(
+    domain: TypedDataDomain,
+    orders: IexecOrders,
+    signers: IexecAccounts,
+): Promise<void> {
+    await signOrder(domain, orders.app, signers.app);
+    if (orders.dataset) {
+        await signOrder(domain, orders.dataset, signers.dataset);
+    }
+    await signOrder(domain, orders.workerpool, signers.workerpool);
+    await signOrder(domain, orders.requester, signers.requester);
 }
 
 /**